rename lex new project components
