+++ src/angular-app/languageforge/lexicon/new-project/new-project-chooser.component.html
