+++ src/angular-app/languageforge/lexicon/new-project/send-receive/new-project-clone.component.html
